+++ src/framework/piral-base/src/loaders/v3/index.ts
@@ -1,14 +1,26 @@
 import { registerDependencyUrls, loadSystemPilet, createEvaluatedPilet } from '../../utils';
 import type { DefaultLoaderConfig, PiletV3Entry, Pilet } from '../../types';
 
+function attachStylesToDocument(pilet: Pilet, url: string) {
+  if (typeof document !== 'undefined') {
+    const link = document.createElement('link');
+    link.setAttribute('data-origin', pilet.name);
+    link.type = 'text/css';
+    link.rel = 'stylesheet';
+    link.href = url;
+    document.head.appendChild(link);
+  }
+}
+
 /**
  * Loads the provided SystemJS-powered pilet.
  * @param entry The pilet's entry.
- * @param _config The loader configuration.
+ * @param options The loader configuration.
  * @returns The evaluated pilet that can now be integrated.
  */
-export default function loader(entry: PiletV3Entry, _config: DefaultLoaderConfig): Promise<Pilet> {
+export default function loader(entry: PiletV3Entry, options: DefaultLoaderConfig): Promise<Pilet> {
   const { dependencies = {}, config = {}, link, ...rest } = entry;
+  const attachStyles = typeof options.attachStyles === 'function' ? options.attachStyles : attachStylesToDocument;
   const meta = {
     dependencies,
     config,
@@ -21,14 +33,9 @@ export default function loader(entry: PiletV3Entry, _config: DefaultLoaderConfig
   return loadSystemPilet(link).then((app) => {
     const pilet = createEvaluatedPilet(meta, app);
 
-    if (Array.isArray(app.styles) && typeof document !== 'undefined') {
+    if (Array.isArray(app.styles)) {
       for (const style of app.styles) {
-        const link = document.createElement('link');
-        link.setAttribute('data-origin', pilet.name);
-        link.type = 'text/css';
-        link.rel = 'stylesheet';
-        link.href = `${pilet.basePath}/${style}`;
-        document.head.appendChild(link);
+        attachStyles(pilet, `${pilet.basePath}/${style}`);
       }
     }