+++ src/framework/piral-base/src/types/lib.ts
@@ -142,4 +142,11 @@ export interface DefaultLoaderConfig {
    * have an integrity defined will be set to "anonymous".
    */
   crossOrigin?: string;
+  /**
+   * Sets the override function for attaching a stylesheet.
+   * This option will only affect `v3` pilets.
+   * @param pilet The pilet containing the style sheet reference.
+   * @param url The style sheet reference URL.
+   */
+  attachStyles?(pilet: Pilet, url: string): void;
 }