+++ src/tooling/piral-cli/src/common/pack.test.ts
@@ -22,25 +22,19 @@ describe('Pack Module', () => {
   it('createPilePackage without package json', async () => {
     json = null;
     expect.assertions(1);
-    await expect(createPiletPackage('./', '', '')).rejects.toEqual(
-      Error('[0020] No valid package.json could be found.'),
-    );
+    await expect(createPiletPackage('./', '', '')).rejects.toThrow('[0020] No valid package.json could be found.');
   });
 
   it('createPilePackage without package json name', async () => {
     json = {};
     expect.assertions(1);
-    await expect(createPiletPackage('./', '', '')).rejects.toEqual(
-      Error('[0021] Cannot pack the pilet - missing name.'),
-    );
+    await expect(createPiletPackage('./', '', '')).rejects.toThrow('[0021] Cannot pack the pilet - missing name.');
   });
 
   it('createPilePackage without package json version', async () => {
     json = { name: 'Test' };
     expect.assertions(1);
-    await expect(createPiletPackage('./', '', '')).rejects.toEqual(
-      Error('[0022] Cannot pack the pilet - missing version.'),
-    );
+    await expect(createPiletPackage('./', '', '')).rejects.toThrow('[0022] Cannot pack the pilet - missing version.');
   });
 
   it('createPilePackage source = target', async () => {